Ignore duplicate phone challenges
Comment out code that stops you from verifying a phone number twice. It isn't really necessary and was slowing Alex down.

My logic was that I didn't want the DB to get into a state where there were multiple verification codes floating around for a phone, but a better solution would be to simply delete/deactivate any previous challenges once a new challenge is created. This is slightly more complex to do and might not even be necessary either.