Add create app and account workflows
This lambda has two mutations:
- createAccount
- createAppAndAccount

The first will take an Athlete with a valid Unit customer ID and request
that Unit create an account. If successful, the account details are
persisted as a new AthleteAccount in TPC. The Unit customer ID is only
created and stored in TPC after a Unit application has been approved.

The second mutation will receive an Athlete and their ssn, and make an
application to Unit. The id of the application is, regardless of the
application's outcome, stored in TPC. If the application is immediately approved
(with Unit says happens for the majority of applications made), this
workflow will go on to create an account for the Athlete as described
above.

In this lambda we talk to two things, the Unit API and our own TPC
graphql API. These side effects are wrapped in src/wrappers. The
environments for these side effects are fetched from env.js.

How I handle environment variables got butchered along the way. I
originally tried to use Reader monads from the monet.js library, but
because my side effects return promises, I wasn't able to bind them
properly. The result is that the environment is hacked in and doesn't
throw errors so that the integration tests can run.

Included are a bunch of integration tests for both workflows, where the
functions exposed in our wrappers are mocked.