Always throw on initial catch of application promise
Js promises are a nightmare. We should be using something like Bluebird
to work with asynchronous code. In this case, I want to store the
application id of a failed Unit application in our DB, but I need to use
catch... which will resolve my error and treat my promise as a success.
So now I need to throw promise rejection inside my catch so my .thens
below dont run.

We just need a tapError.